AntiScamBot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
cryptography==41.0.7
python-multipart==0.0.6
pytz==2023.3